+++ src/main/java/com/example/springserver/domain/caregiver/controller/CaregiverController.java
@@ -1,9 +1,11 @@
 package com.example.springserver.domain.caregiver.controller;
 
 import com.example.springserver.domain.caregiver.converter.CaregiverConverter;
-import com.example.springserver.domain.caregiver.dto.response.CaregiverResponseDTO;
+import com.example.springserver.domain.caregiver.dto.request.CaregiverRequestDto;
 import com.example.springserver.domain.caregiver.entity.Caregiver;
 import com.example.springserver.domain.caregiver.service.CareGiverService;
+import com.example.springserver.domain.match.dto.request.MatchRequestDto;
+import com.example.springserver.domain.match.dto.request.MatchRequestDto.RecruitReq;
 import com.example.springserver.global.security.util.CustomUserDetails;
 import com.example.springserver.service.JoinService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -14,10 +16,9 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 
-import static com.example.springserver.domain.caregiver.dto.request.CaregiverRequestDTO.*;
-import static com.example.springserver.domain.caregiver.dto.response.CaregiverResponseDTO.*;
+import static com.example.springserver.domain.caregiver.dto.request.CaregiverRequestDto.*;
+import static com.example.springserver.domain.caregiver.dto.response.CaregiverResponseDto.*;
 
 @RestController
 @Validated
@@ -31,7 +32,7 @@ public class CaregiverController {
 
     @Operation(summary = "회원가입", description = "Post")
     @PostMapping("/signup")
-    public SignUpCaregiverResult signUpCaregiver(@RequestBody @Valid SignUpCaregiverReq request){
+    public SignUpCaregiverResult signUpCaregiver(@RequestBody @Valid CaregiverRequestDto.SignUpCaregiverReqDto request){
         Caregiver newCaregiver = joinService.signUpCaregiver(request);
         return CaregiverConverter.toSignUpCaregiverResult(newCaregiver);
     }
@@ -46,7 +47,7 @@ public class CaregiverController {
     @Operation(summary = "요양보호사 정보수정", description = "Put")
     @PutMapping("/profile")
     public CareGiverInfoResponseDTO updateCaregiver(@AuthenticationPrincipal CustomUserDetails user,
-                                                    @RequestBody @Valid UpdateCaregiverReq request){
+                                                    @RequestBody @Valid CaregiverRequestDto.UpdateCaregiverReqDto request){
         Caregiver searched = careGiverService.updateUserInfo(user,request);
         return CaregiverConverter.infoResponseDto(searched);
     }
@@ -56,52 +57,4 @@ public class CaregiverController {
     public Boolean changeStatus(@AuthenticationPrincipal CustomUserDetails user){
         return careGiverService.changeStatus(user);
     }
-
-    @Operation(summary = "요양보호사 구직조건등록 및 수정", description = "Post")
-    @PostMapping("/jobcondition")
-    public JobConditionResponseDTO createJobCondition(@AuthenticationPrincipal CustomUserDetails user,
-                                                      @RequestBody @Valid JobConditionRequestDTO request){
-        return  careGiverService.createOrUpdateJobCondition(user,request);
-    }
-
-    @Operation(summary = "요양보호사 구직조건수정", description = "Put")
-    @PutMapping("/jobcondition")
-    public JobConditionResponseDTO updateJobCondition(@AuthenticationPrincipal CustomUserDetails user,
-                                                      @RequestBody @Valid JobConditionRequestDTO request){
-        return  careGiverService.createOrUpdateJobCondition(user,request);
-    }
-
-    @Operation(summary = "요양보호사 구직정보조회", description = "Get")
-    @GetMapping("/jobcondition")
-    public JobConditionResponseDTO getJobCondition(@AuthenticationPrincipal CustomUserDetails user){
-        return careGiverService.getJobCondition(user);
-    }
-
-    @Operation(summary = "요양보호사 상세정보조회", description = "Get")
-    @GetMapping("/detail")
-    public DetailJobConditionResponseDTO getDetailJobCondition(@AuthenticationPrincipal CustomUserDetails user){
-        return careGiverService.getDetailedJobCondition(user);
-    }
-
-    @Operation(summary = "요양보호사 구인요청응답", description = "Put")
-    @PutMapping("/response")
-    public String responseToRecruit(@AuthenticationPrincipal CustomUserDetails user,
-                                                      @RequestBody @Valid RecruitReq request){
-        return careGiverService.responseToRecruit(user,request);
-    }
-
-    @Operation(summary = "요양보호사 매칭현황리스트 조회", description = "Get")
-    @GetMapping("/matching")
-    public MatchedListRes responseToRecruit(@AuthenticationPrincipal CustomUserDetails user){
-        return CaregiverConverter.toMatchedListRes(careGiverService.getCalenderList(user));
-    }
-
-
-    @Operation(summary = "요양보호사 근무요청리스트 조회", description = "Get")
-    @GetMapping("/requests")
-    public RequestsListRes getListOfRequests(@AuthenticationPrincipal CustomUserDetails user){
-        return CaregiverConverter.toRequestListRes(careGiverService.getRequests(user));
-    }
-
-
 }
