+++ src/main/java/com/example/springserver/domain/caregiver/dto/response/CaregiverResponseDto.java
@@ -4,7 +4,6 @@ import com.example.springserver.domain.caregiver.entity.enums.CertType;
 import com.example.springserver.domain.caregiver.entity.enums.Level;
 import com.example.springserver.domain.caregiver.entity.enums.ScheduleAvailability;
 import com.example.springserver.domain.caregiver.entity.enums.Sexual;
-import com.example.springserver.domain.center.entity.RecruitTime;
 import com.example.springserver.domain.center.entity.enums.CareType;
 import com.example.springserver.domain.center.entity.enums.ElderRate;
 import com.example.springserver.domain.center.entity.enums.Week;
@@ -15,7 +14,7 @@ import lombok.NoArgsConstructor;
 
 import java.util.List;
 
-public class CaregiverResponseDTO {
+public class CaregiverResponseDto {
 
     @Builder
     @Getter
@@ -51,140 +50,6 @@ public class CaregiverResponseDTO {
         private List<ExperienceResponseDTO> experienceResponseDTOList ;
     }
 
-    @Getter
-    @Builder
-    public static class JobConditionResponseDTO {
-
-        private Long jobConditionId;
-
-        private ScheduleAvailability flexibleSchedule;
-
-        private Integer desiredHourlyWage;
-
-        private ScheduleAvailability selfFeeding;
-
-        private ScheduleAvailability mealPreparation;
-
-        private ScheduleAvailability cookingAssistance;
-
-        private ScheduleAvailability enteralNutritionSupport;
-
-        private ScheduleAvailability selfToileting;
-
-        private ScheduleAvailability occasionalToiletingAssist;
-
-        private ScheduleAvailability diaperCare;
-
-        private ScheduleAvailability catheterOrStomaCare;
-
-        private ScheduleAvailability independentMobility;
-
-        private ScheduleAvailability mobilityAssist;
-
-        private ScheduleAvailability wheelchairAssist;
-
-        private ScheduleAvailability immobile;
-
-        private ScheduleAvailability cleaningLaundryAssist;
-
-        private ScheduleAvailability bathingAssist;
-
-        private ScheduleAvailability hospitalAccompaniment;
-
-        private ScheduleAvailability exerciseSupport;
-
-        private ScheduleAvailability emotionalSupport;
-
-        private ScheduleAvailability cognitiveStimulation;
-
-        private String dayOfWeek;
-
-        private Long startTime;
-
-        private Long endTime;
-
-        private List<LocationResponseDTO> locationRequestDTOList;
-    }
-
-    @Getter
-    @Builder
-    public static class DetailJobConditionResponseDTO {
-
-        //CareGiver info
-
-        private String name;
-
-        private String contact;
-
-        private Boolean car;
-
-        private Boolean education;
-
-        private String img;
-
-        private String intro;
-
-        private String address;
-
-        private Boolean employmentStatus;
-
-        private List<CertificateResponseDTO> certificateResponseDTOList;
-
-        private List<ExperienceResponseDTO> experienceResponseDTOList ;
-
-        //jobCondition
-
-        private Long jobConditionId;
-
-        private ScheduleAvailability flexibleSchedule;
-
-        private Integer desiredHourlyWage;
-
-        private ScheduleAvailability selfFeeding;
-
-        private ScheduleAvailability mealPreparation;
-
-        private ScheduleAvailability cookingAssistance;
-
-        private ScheduleAvailability enteralNutritionSupport;
-
-        private ScheduleAvailability selfToileting;
-
-        private ScheduleAvailability occasionalToiletingAssist;
-
-        private ScheduleAvailability diaperCare;
-
-        private ScheduleAvailability catheterOrStomaCare;
-
-        private ScheduleAvailability independentMobility;
-
-        private ScheduleAvailability mobilityAssist;
-
-        private ScheduleAvailability wheelchairAssist;
-
-        private ScheduleAvailability immobile;
-
-        private ScheduleAvailability cleaningLaundryAssist;
-
-        private ScheduleAvailability bathingAssist;
-
-        private ScheduleAvailability hospitalAccompaniment;
-
-        private ScheduleAvailability exerciseSupport;
-
-        private ScheduleAvailability emotionalSupport;
-
-        private ScheduleAvailability cognitiveStimulation;
-
-        private String dayOfWeek;
-
-        private Long startTime;
-
-        private Long endTime;
-
-        private List<LocationResponseDTO> locationRequestDTOList;
-
-    }
 
     @Getter
     @Builder
