+++ src/main/java/com/example/springserver/domain/caregiver/service/CommonService.java
@@ -1,0 +1,22 @@
+package com.example.springserver.domain.caregiver.service;
+
+import com.example.springserver.domain.caregiver.entity.Caregiver;
+import com.example.springserver.domain.caregiver.repository.CaregiverRepository;
+import com.example.springserver.global.apiPayload.format.ErrorCode;
+import com.example.springserver.global.apiPayload.format.GlobalException;
+import com.example.springserver.global.security.util.CustomUserDetails;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class CommonService {
+
+    private final CaregiverRepository caregiverRepository;
+
+    public Caregiver getById(CustomUserDetails user) throws GlobalException {
+        return caregiverRepository.findById(user.getId()).orElseThrow(()-> new GlobalException(ErrorCode.USER_NOT_FOUND));
+    }
+}
