+++ src/main/java/com/example/springserver/domain/match/dto/request/MatchRequestDto.java
@@ -1,11 +1,10 @@
-package com.example.springserver.domain.caregiver.dto.request;
+package com.example.springserver.domain.match.dto.request;
 
 import com.example.springserver.domain.caregiver.entity.enums.CertType;
 import com.example.springserver.domain.caregiver.entity.enums.Level;
 import com.example.springserver.domain.caregiver.entity.enums.ScheduleAvailability;
 import com.example.springserver.domain.center.entity.enums.RecruitStatus;
 import com.example.springserver.domain.center.entity.enums.Week;
-import jakarta.persistence.Column;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
@@ -13,78 +12,7 @@ import lombok.Getter;
 
 import java.util.List;
 
-public class CaregiverRequestDTO {
-
-    @Getter
-    public static class SignUpCaregiverReq{
-
-        @NotEmpty
-        private String username;
-
-        @NotEmpty
-        private String password;
-
-        @NotEmpty
-        private String name;
-
-        @NotEmpty
-        private String contact;
-
-        @NotNull
-        private Boolean car;
-
-        @NotNull
-        private Boolean education;
-
-        private String img;
-
-        private String intro;
-
-        private String address;
-
-        private Boolean employmentStatus;
-
-        private List<CertificateRequestDTO> certificateRequestDTOList;
-
-        private List<ExperienceRequestDTO> experienceRequestDTOList;
-
-
-        public SignUpCaregiverReq() {
-            this.car = false; // 기본값 false
-            this.education = false;   // 기본값 false
-        }
-
-        public void setCommonImg(){
-            this.img = "http://localhost:8080/basicImg.jpeg";
-        }
-    }
-
-    @Getter
-    public static class UpdateCaregiverReq{
-
-        @NotEmpty
-        private String username;
-
-        @NotEmpty
-        private String contact;
-
-        @NotNull
-        private Boolean car;
-
-        @NotNull
-        private Boolean education;
-
-        private String img;
-
-        private String intro;
-
-        private String address;
-
-        private List<CertificateRequestDTO> certificateRequestDTOList;
-
-        private List<ExperienceRequestDTO> experienceRequestDTOList;
-
-    }
+public class MatchRequestDto {
 
     @Getter
     public static class JobConditionRequestDTO {
@@ -227,31 +155,4 @@ public class CaregiverRequestDTO {
         private Long endTime;
     }
 
-
-    @Getter
-    public static class CertificateRequestDTO{
-
-        private String certNum;
-
-        @NotNull
-        private CertType certType;
-
-        @NotNull
-        private Level certRate;
-
-    }
-
-    @Getter
-    public static class ExperienceRequestDTO{
-
-        @NotNull
-        private int duration;
-
-        @NotEmpty
-        private String title;
-
-        @NotEmpty
-        private String description;
-
-    }
 }
