+++ src/main/resources/application.yml
@@ -1,6 +1,6 @@
 spring:
   profiles:
-    active: prod
+    active: local
 server:
   port: 8080
   address: 0.0.0.0
